refactor: mv profile components
