+++ src/components/Profile/ContentCollected.tsx
