+++ src/components/Profile/FarcasterProfile.tsx
@@ -2,11 +2,11 @@ import { useQuery } from '@tanstack/react-query';
 import { useMemo } from 'react';
 import { useDocumentTitle } from 'usehooks-ts';
 
-import ContentTabs from '@/app/(normal)/profile/components/ContentTabs.js';
-import Info from '@/app/(normal)/profile/components/Info.js';
-import Title from '@/app/(normal)/profile/components/Title.js';
 import Loading from '@/components/Loading.js';
 import NotFoundFallback from '@/components/NotFoundFallback.js';
+import ContentTabs from '@/components/Profile/ContentTabs.js';
+import Info from '@/components/Profile/Info.js';
+import Title from '@/components/Profile/Title.js';
 import { createPageTitle } from '@/helpers/createPageTitle.js';
 import { useLogin } from '@/hooks/useLogin.js';
 import { usePlatformAccount } from '@/hooks/usePlatformAccount.js';
